Add (failing) test about preserving comment order
Some doxygen comments are not associated with any element translatable
into C++, e.g. opening and closing groups: those are stand-alone.
Accordingly, comments (and the elements they may refer to) must not be
resorted when generating the output.

In addition, one-line comments must preserve empty lines between them:
putting them together causes the interpretation to change, e.g. a @{
followed by a @brief always opens a group, but the brief is part of the
group instead of the following element if there is no empty line between
the two comments.

This test case show-cases the problem without having an implementation,
yet.