Stop sorting class members by access
Output class members in the order they were defined in, switching
between public and private access as needed. This is needed as a
preparation for getting the comment-order functional test to pass: it
free-standing comment blocks that must be kept in their original order.

As a side effect, the output no longer contains an obligatory "public:"
access-declaration at the start of the class. This requires changing
some of the existing test cases with empty classes or where the first
element is private.