Restore output for siblings of internal components
Since doxyqml!16, commit 371b65ac, doxyqml no longer generates any
output for any classes declared in the same `qmldir` file as an internal
component or even a public component with the substring `internal` in
its name.

Define two functional tests for this and fix the problem. Also change
the return value of the main-function for an internal file to 0 so that
the functional test still works.

Refs BUG:488682
Fixes #16